gcl/gcl/requirements.txt
cloudpickle==1.6.0
cycler==0.10.0
future==0.18.2
gym==0.18.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.1
Pillow==7.2.0
protobuf==3.15.3
pybullet==3.0.8
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.0
six==1.15.0
stable-baselines3==0.10.0
tensorboardX==2.1
typing-extensions==3.7.4.3
